change BATTERY_CHANGED action extra to EXTRA_PLUGGED
Previously, users on the Android default emulator had to run
power status discharging and power ac off to get the static
broadcastreceiver to work. Now both static and dynamic
broadcasts will work with power ac off/on in telnet.

We are now checking whether an AC adapter is connected, as
opposed to checking whether we're charging or not (AC or USB)